# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /ads:
    get:
      tags:
      - Объявления
      summary: Получение всех объявлений
      operationId: getAllAds
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdsDto'
    post:
      tags:
      - Объявления
      summary: Добавить объявление
      operationId: addAd
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateAdDto'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdDto'
  /ads/{id}/comments:
    get:
      tags:
      - Комментарии
      summary: Получить комментарии объявления
      operationId: getComments
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentsDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentsDto'
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentsDto'
    post:
      tags:
      - Комментарии
      summary: Добавить комментарий к объявлению
      operationId: addComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateCommentDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
...
  /ads/{id}:
    get:
      tags:
      - Объявления
      summary: Получить информацию об объявлении
      operationId: getAds
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExtendedAdDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExtendedAdDto'
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ExtendedAdDto'
    delete:
      tags:
      - Объявления
      summary: Удалить объявление
      operationId: removeAd
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      tags:
      - Объявления
      summary: Обновить информацию об объявлении
      operationId: updateAds
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateAdDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdDto'
  /ads/{id}/image:
    patch:
      tags:
      - Объявления
      summary: Обновить картинку объявления
      operationId: updateImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/octet-stream: {}
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
                      format: byte
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
                      format: byte
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
                      format: byte
  /ads/{adId}/comments/{commentId}:
    delete:
      tags:
      - Комментарии
      summary: Удалить комментарий
      operationId: deleteComment
      parameters:
      - name: adId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "204":
          description: No Content
    patch:
      tags:
      - Комментарии
      summary: Обновить комментарий
      operationId: updateComment
      parameters:
      - name: adId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateCommentDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommentDto'
...
  /ads/me:
    get:
      tags:
      - Объявления
      summary: Получить объявления авторизованного пользователя
      operationId: getAdsMe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdsDto'
  /ads/image/{id}/from-db:
    get:
      tags:
      - ads-controller
      operationId: getAdImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
                      format: byte
components:
  schemas:
...
    CreateOrUpdateAdDto:
      type: object
      properties:
        title:
          maxLength: 32
          minLength: 4
          type: string
          description: заголовок объявления
        price:
          type: integer
          description: цена объявления
          format: int32
        description:
          maxLength: 64
          minLength: 8
          type: string
          description: описание объявления
    AdDto:
      type: object
      properties:
        author:
          type: integer
          description: id автора объявления
          format: int32
        image:
          type: string
          description: ссылка на картинку объявления
        pk:
          type: integer
          description: id объявления
          format: int32
        price:
          type: integer
          description: цена объявления
          format: int32
        title:
          type: string
          description: заголовок объявления
    CreateOrUpdateCommentDto:
      type: object
      properties:
        text:
          maxLength: 64
          minLength: 8
          type: string
          description: текст комментария
    CommentDto:
      type: object
      properties:
        author:
          type: integer
          description: id автора комментария
          format: int32
        authorImage:
          type: string
          description: ссылка на аватар автора комментария
        authorFirstName:
          type: string
          description: имя создателя комментария
        createdAt:
          type: integer
          description: дата и время создания комментария в миллисекундах с 00:00:00
            01.01.1970
          format: int64
        pk:
          type: integer
          description: id комментария
          format: int32
        text:
          type: string
          description: текст комментария
...
    AdsDto:
      type: object
      properties:
        count:
          type: integer
          description: общее количество объявлений
          format: int32
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdDto'
    ExtendedAdDto:
      type: object
      properties:
        pk:
          type: integer
          description: id объявления
          format: int32
        authorFirstName:
          type: string
          description: имя автора объявления
        authorLastName:
          type: string
          description: фамилия автора объявления
        description:
          type: string
          description: описание объявления
        email:
          type: string
          description: логин автора объявления
        image:
          type: string
          description: ссылка на картинку объявления
        phone:
          type: string
          description: телефон автора объявления
        price:
          type: integer
          description: цена объявления
          format: int32
        title:
          type: string
          description: заголовок объявления
    CommentsDto:
      type: object
      properties:
        count:
          type: integer
          description: общее количество комментариев
          format: int32
        results:
          type: array
          items:
            $ref: '#/components/schemas/CommentDto'