# src/main/java/ru/skypro/homework/dto/AdsDto.java
package ru.skypro.homework.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

import java.util.List;

@Data
public class AdsDto {
    @Schema(description = "общее количество объявлений")
    private int count;
    private List<AdDto> results;

}