# src/main/java/ru/skypro/homework/dto/CreateOrUpdateCommentDto.java
package ru.skypro.homework.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

import javax.validation.constraints.Size;

@Data
public class CreateOrUpdateCommentDto {
    @Schema(description = "текст комментария")
    @Size(min = 8, max = 64)
    private String text;
}